refactor(kubernetes): move supervisor bootstrap identity into driver
