+++ crates/openshell-core/src/lib.rs
@@ -37,6 +37,7 @@ pub mod provider_credentials;
 pub mod sandbox_env;
 pub mod secrets;
 pub mod settings;
+pub mod supervisor_bootstrap;
 pub mod telemetry;
 pub mod time;
 pub mod transport_errors;
