+++ crates/openshell-driver-kubernetes/src/lib.rs
@@ -1,10 +1,14 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod bootstrap;
 pub mod config;
 pub mod driver;
 pub mod grpc;
 
+pub use bootstrap::{
+    K8sIdentityResolver, KubernetesSupervisorBootstrapIdentityProvider, LiveK8sResolver,
+};
 pub use config::{
     AppArmorProfile, DEFAULT_PROXY_UID, DEFAULT_SANDBOX_SERVICE_ACCOUNT_NAME,
     DEFAULT_WORKSPACE_STORAGE_SIZE, KubernetesComputeConfig, KubernetesSidecarConfig,
