+++ crates/openshell-server/src/auth/authenticator.rs
@@ -14,8 +14,8 @@
 //!
 //! Live authenticators slotting into the chain:
 //! - [`super::sandbox_jwt::SandboxJwtAuthenticator`] — gateway-minted JWTs
-//! - [`super::k8s_sa::K8sServiceAccountAuthenticator`] — K8s projected SA
-//!   tokens (path-scoped to Kubernetes bootstrap RPCs)
+//! - [`super::k8s_sa::SupervisorBootstrapAuthenticator`] — driver-provided
+//!   bootstrap tokens (path-scoped to supervisor bootstrap RPCs)
 //! - [`super::oidc::OidcAuthenticator`] — user OIDC Bearer tokens
 use super::principal::Principal;
 use async_trait::async_trait;
